pyproject.toml
[project]
name = "logsit-agent"
version = "0.1.0"
description = "AI agent for logging and conversation management with MCP support"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Logsit Team", email = "team@logsit.ai"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.12.5,<3.0.0",
    "fastapi>=0.135.3,<1.0.0",
    "uvicorn>=0.44.0,<1.0.0",
    "openai>=2.31.0,<3.0.0",
    "aiofiles>=24.1.0,<25.0.0",
    "python-dotenv>=1.2.2,<2.0.0",
    "rich>=14.3.3,<15.0.0",
    "tenacity>=9.1.4,<10.0.0",
]

[project.optional-dependencies]
mcp = [
    "mcp>=1.27.0,<2.0.0",
    "protobuf>=7.34.1,<8.0.0",
]
rag = [
    "sentence-transformers>=3.4.1,<4.0.0",
    "faiss-cpu>=1.9.0,<2.0.0",
    "numpy>=2.2.6,<3.0.0",
    "pandas>=2.3.3,<3.0.0",
]
ui = [
    "gradio>=6.11.0,<7.0.0",
    "streamlit>=1.56.0,<2.0.0",
]
dev = [
    "pytest>=7.0.0,<8.0.0",
    "pytest-asyncio>=0.21.0,<1.0.0",
    "httpx>=0.28.1,<1.0.0",
    "black>=23.0.0,<24.0.0",
    "ruff>=0.4.0,<0.5.0",
    "mypy>=1.8.0,<2.0.0",
]
full = [
    "logsit-agent[mcp,rag,ui,dev]",
]

[project.urls]
Homepage = "https://github.com/yourusername/logsit-agent"
Documentation = "https://logsit-agent.readthedocs.io"
Repository = "https://github.com/yourusername/logsit-agent"
Issues = "https://github.com/yourusername/logsit-agent/issues"

[project.scripts]
logsit-agent = "app.main:run"
logsit-cli = "cli.main:cli"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "core*", "services*", "utils*"]

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff]
target-version = "py39"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
]
exclude = [
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".tox",
    ".venv",
    "_build",
    "buck-out",
    "build",
    "dist",
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true